Updates
Minor changes

+++ ScalingButton.java
@@ -22,7 +22,7 @@ public class ScalingButton extends JButton implements ComponentListener{
 		Font font = getFont();
 		//had to do by height because words are different lengths and don't want different font sizes
 		float size = getHeight()/4; //this was a guess and check thing I still don't know what font size is based on
-		setFont(font.deriveFont(font.getStyle(),size));
+		setFont(font.deriveFont(font.getStyle(),(int)size));
 	}
 	@Override
 	public void componentShown(ComponentEvent arg0) {
